+++ examples/templates/suite_com/test_com.py
@@ -1,5 +1,5 @@
 ##########################################################################
-# Copyright (c) 2010-2021 Robert Bosch GmbH
+# Copyright (c) 2010-2022 Robert Bosch GmbH
 # This program and the accompanying materials are made available under the
 # terms of the Eclipse Public License 2.0 which is available at
 # http://www.eclipse.org/legal/epl-2.0.
@@ -24,6 +24,7 @@ import pykiso
 
 # as usual import your auxiliairies
 from pykiso.auxiliaries import com_aux
+from pykiso.lib.connectors.cc_raw_loopback import CCLoopback
 
 
 @pykiso.define_test_parameters(
@@ -38,7 +39,20 @@ class TestCaseOverride(pykiso.BasicTest):
 
     def setUp(self):
         """If a fixture is not use just override it like below."""
-        pass
+        logging.info(
+            f"--------------- SETUP: {self.test_suite_id}, {self.test_case_id} ---------------"
+        )
+        # just create a communication auxiliary with a bran new channel
+        # in order to send odd bytes. Always use named arguments!
+        com_aux.start()
+        self.com_aux_odd = com_aux.create_copy(com=CCLoopback())
+        # just create a communication auxiliary based on the given
+        # parameters present in the yaml config (com_aux.yaml) to send
+        # even bytes
+        self.com_aux_even = self.com_aux_odd.create_copy()
+        # start com_aux_even because original configuration has
+        # auto_start flag set to False
+        self.com_aux_even.start()
 
     def test_run(self):
         """Thanks to the usage of dev cc_raw_loopback, let's try to send
@@ -47,14 +61,32 @@ class TestCaseOverride(pykiso.BasicTest):
         logging.info(
             f"--------------- RUN: {self.test_suite_id}, {self.test_case_id} ---------------"
         )
-        # send a some bytes
-        com_aux.send_message(b"\x01\x02\x03")
-        # receive some bytes and check if it was our previous sent
-        # message
+        # first send the even bytes and stop the copy in order to
+        # automatically resume the original one (self.com_aux_odd)
+        self.com_aux_even.send_message(b"\x02\x04\x06")
+        response = self.com_aux_even.receive_message()
+        self.assertEqual(response, b"\x02\x04\x06")
+        # destroy com_aux_even and start working with com_aux_odd aux
+        # copy
+        self.com_aux_odd.destroy_copy()
+
+        # Then send the odd bytes and stop the copy in order to
+        # automatically resume the original one (com_aux)
+        self.com_aux_odd.send_message(b"\x01\x03\x05")
+        response = self.com_aux_odd.receive_message()
+        self.assertEqual(response, b"\x01\x03\x05")
+        # destroy com_aux_odd and start working with com_aux
+        com_aux.destroy_copy()
+
+        # Just use the original communication auxiliary to send and
+        # receive some bytes
+        com_aux.send_message(b"\x01\x02\x03\x04\x05\x06")
         response = com_aux.receive_message()
         logging.info(f"received message: {response}")
-        self.assertEqual(response, b"\x01\x02\x03")
+        self.assertEqual(response, b"\x01\x02\x03\x04\x05\x06")
 
     def tearDown(self):
         """If a fixture is not use just override it like below."""
-        pass
+        logging.info(
+            f"--------------- TEARDOWN: {self.test_suite_id}, {self.test_case_id} ---------------"
+        )