+++ src/pykiso/lib/auxiliaries/example_test_auxiliary.py
@@ -1,5 +1,5 @@
 ##########################################################################
-# Copyright (c) 2010-2021 Robert Bosch GmbH
+# Copyright (c) 2010-2022 Robert Bosch GmbH
 # This program and the accompanying materials are made available under the
 # terms of the Eclipse Public License 2.0 which is available at
 # http://www.eclipse.org/legal/epl-2.0.
@@ -38,7 +38,7 @@ class ExampleAuxiliary(AuxiliaryInterface):
         com=None,
         flash=None,
         **kwargs,
-    ):
+    ) -> None:
         """Constructor.
 
         :param name: Alias of the auxiliary instance