+++ src/pykiso/lib/connectors/cc_socket_can/__init__.py
@@ -1,0 +1,20 @@
+##########################################################################
+# Copyright (c) 2010-2022 Robert Bosch GmbH
+# This program and the accompanying materials are made available under the
+# terms of the Eclipse Public License 2.0 which is available at
+# http://www.eclipse.org/legal/epl-2.0.
+#
+# SPDX-License-Identifier: EPL-2.0
+##########################################################################
+
+"""
+socketCan
+*********
+
+:module: cc_socket_can
+
+.. currentmodule:: cc_socket_can
+
+"""
+
+from .cc_socket_can import CCSocketCan