+++ src/pykiso/lib/robot_framework/acroname_auxiliary.py
@@ -1,5 +1,5 @@
 ##########################################################################
-# Copyright (c) 2010-2021 Robert Bosch GmbH
+# Copyright (c) 2010-2022 Robert Bosch GmbH
 # This program and the accompanying materials are made available under the
 # terms of the Eclipse Public License 2.0 which is available at
 # http://www.eclipse.org/legal/epl-2.0.
@@ -49,7 +49,7 @@ class AcronameAuxiliary(RobotAuxInterface):
         aux = self._get_aux(aux_alias)
         return aux.set_port_enable(port)
 
-    @keyword(name="Set Port disable")
+    @keyword(name="Set port disable")
     def set_port_disable(self, aux_alias: str, port: int) -> int:
         """Disable power and data lines for a USB port.