+++ src/pykiso/test_setup/config_registry.py
@@ -1,5 +1,5 @@
 ##########################################################################
-# Copyright (c) 2010-2021 Robert Bosch GmbH
+# Copyright (c) 2010-2022 Robert Bosch GmbH
 # This program and the accompanying materials are made available under the
 # terms of the Eclipse Public License 2.0 which is available at
 # http://www.eclipse.org/legal/epl-2.0.
@@ -79,6 +79,17 @@ class ConfigRegistry:
             for alias, inst in all_auxes.items()
             if isinstance(inst, aux_type)
         }
+
+    @classmethod
+    def get_aux_config(cls, name: str) -> dict:
+        """Return the registered auxiliary configuration based on his
+        name.
+
+        :param name: auxiliary alias
+
+        :return: auxiliary's configuration (yaml content)
+        """
+        return ConfigRegistry._linker._aux_cache.configs[name]
 
     @classmethod
     def get_auxes_alias(cls) -> list: