+++ tests/conftest.py
@@ -1,5 +1,5 @@
 ##########################################################################
-# Copyright (c) 2010-2021 Robert Bosch GmbH
+# Copyright (c) 2010-2022 Robert Bosch GmbH
 # This program and the accompanying materials are made available under the
 # terms of the Eclipse Public License 2.0 which is available at
 # http://www.eclipse.org/legal/epl-2.0.
@@ -148,6 +148,7 @@ def CustomTestCaseAndSuite(request):
             for _ in range(max_com):
                 self.connectors.append(self.channel_in_use())
                 self.auxiliaries.append(self.auxiliary_in_use(com=self.connectors[-1]))
+                self.auxiliaries[-1].start()
                 self.auxiliaries[-1].create_instance()
 
         def prepare_default_test_cases(self, param):